fix(nfts): prevent asset page events updates
