+++ packages/blockchain-wallet-v4-frontend/src/scenes/Nfts/Asset/components/index.tsx
@@ -1,0 +1,19 @@
+import { colors } from '@blockchain-com/constellation'
+import styled from 'styled-components'
+
+import { Text } from 'blockchain-info-components'
+
+export const Highest = styled(Text)`
+  margin-bottom: 12px;
+  font-weight: 600;
+  font-size: 14px;
+  color: ${colors.grey600};
+`
+
+export const EthText = styled(Highest)`
+  font-size: 24px;
+  display: flex;
+  margin-bottom: 20px;
+  align-items: center;
+  color: ${colors.grey900};
+`
